pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "site2skillgo"
version = "0.1.0"
description = "Turn any website into a Claude Skill"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.11.0",
    "pyyaml>=6.0",
]

[project.scripts]
site2skillgo = "site2skillgo.main:main"

[tool.setuptools]
packages = ["site2skillgo"]
package-dir = {"" = "."}

[tool.setuptools.package-data]
site2skillgo = ["templates/*"]

[tool.uv]
# No special uv settings needed yet
